Swap FILTER_SANITIZE_STRING to FILTER_UNSAFE_RAW
The FILTER_SANITIZE_STRING filter is deprecated as of PHP 8.1

Reviewing this code, it was determined the $nonce variable is used
exclusively as an argument for wp_verify_nonce(), which does not
require a filtered value.

Based on this, we're now using the FILTER_UNSAFE_RAW filter, which makes
no changes to the value but allows us to keep using the filter_input()
function here (like we do for other values).

Resolves #2